Refactor coordination service package
