+++ ydb/aio/__init__.py
@@ -1,4 +1,5 @@
 from .driver import Driver  # noqa
 from .table import SessionPool, retry_operation  # noqa
 from .query import QuerySessionPool, QuerySession, QueryTxContext  # noqa
-from .coordination_client import CoordinationClient  # noqa
+
+# from .coordination import CoordinationClient  # noqa
