+++ ydb/aio/coordination/__init__.py
@@ -1,0 +1,5 @@
+__all__ = [
+    "CoordinationClient",
+]
+
+from .client import CoordinationClient
